Refactor to rename package model to domain and extract score from domain
