+++ optimizer/src/main/java/es/carloscasalar/nwp/domain/Character.java
@@ -1,4 +1,4 @@
-package es.carloscasalar.nwp.model;
+package es.carloscasalar.nwp.domain;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
