+++ optimizer/src/main/java/es/carloscasalar/nwp/rest/PlanController.java
@@ -1,7 +1,7 @@
 package es.carloscasalar.nwp.rest;
 
-import es.carloscasalar.nwp.model.Plan;
-import es.carloscasalar.nwp.model.PlanRequest;
+import es.carloscasalar.nwp.domain.Plan;
+import es.carloscasalar.nwp.domain.PlanRequest;
 import es.carloscasalar.nwp.usecase.GeneratePlan;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
