+++ optimizer/src/test/java/es/carloscasalar/nwp/score/WatchDifficultyWeightComparatorTest.java
@@ -1,7 +1,7 @@
-package es.carloscasalar.nwp.model.score;
+package es.carloscasalar.nwp.score;
 
-import es.carloscasalar.nwp.model.Watch;
-import es.carloscasalar.nwp.model.fixtures.CharacterFactory;
+import es.carloscasalar.nwp.domain.Watch;
+import es.carloscasalar.nwp.domain.fixtures.CharacterFactory;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
